+++ ile_de_france/src/main/java/org/eqasim/ile_de_france/RunSimulation.java
@@ -1,5 +1,6 @@
 package org.eqasim.ile_de_france;
 
+import org.eqasim.core.simulation.analysis.EqasimAnalysisModule;
 import org.eqasim.core.simulation.mode_choice.EqasimModeChoiceModule;
 import org.eqasim.ile_de_france.mode_choice.IDFModeChoiceModule;
 import org.matsim.api.core.v01.Scenario;
@@ -27,6 +28,7 @@ public class RunSimulation {
 
 		Controler controller = new Controler(scenario);
 		IDFConfigurator.configureController(controller);
+		controller.addOverridingModule(new EqasimAnalysisModule());
 		controller.addOverridingModule(new EqasimModeChoiceModule());
 		controller.addOverridingModule(new IDFModeChoiceModule(cmd));
 		controller.run();